Add TDP to the list of apps to be converted
The cfgov-refresh satellite app teachers_digital_platform doesn't have any Django models yet, but may soon. In an attempt to be forward looking, this change adds TDP to the list of tables that are converted from MySQL to Postgres by this project. If no tables exist, this change won't make any difference. If some are added, they will be converted.

If some tables were to be added without this change, they wouldn't be included in Postgres converted dumps, and thus could cause problems downstream.